requirements.txt
streamlit>=1.32.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
sentence-transformers>=2.5.1
transformers>=4.40.0
faiss-cpu>=1.7.4
pymupdf>=1.23.5
python-docx>=1.1.0
pyyaml>=6.0.1
tqdm>=4.66.1
python-dotenv>=1.0.1
beautifulsoup4>=4.12.2
lxml>=5.1.0
python-pptx>=0.6.21
